% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_amino_acids}
\alias{as_amino_acids}
\alias{as_amino_acids,AAString-method}
\alias{as_amino_acids,DNAString-method}
\alias{as_amino_acids,RNAString-method}
\title{Amino Acid Conversion}
\usage{
as_amino_acids(sequence)

\S4method{as_amino_acids}{AAString}(sequence)

\S4method{as_amino_acids}{DNAString}(sequence)

\S4method{as_amino_acids}{RNAString}(sequence)
}
\arguments{
\item{sequence}{an object representing an amino acid sequence}
}
\value{
a character vector of amino acids
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Method for converting from biological strings in various formats
to an amino acid sequence.
}
\details{
This converts a sequence to a character vector of amino acids.
This function is not vectorized. Supported input types are \code{DNAString},
\code{RNAString}, and \code{character} vector.
}
\examples{
orf <- Biostrings::AAString("MERVIN*")
as_amino_acids(orf)

orf <- Biostrings::DNAString("ATGAAATAG")
as_amino_acids(orf)


orf <- Biostrings::RNAString("AUGAAAUAG")
as_amino_acids(orf)

}
