% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_codons}
\alias{as_codons}
\alias{as_codons,DNAString-method}
\alias{as_codons,RNAString-method}
\alias{as_codons,character-method}
\title{Codon Conversion}
\usage{
as_codons(sequence)

\S4method{as_codons}{DNAString}(sequence)

\S4method{as_codons}{RNAString}(sequence)

\S4method{as_codons}{character}(sequence)
}
\arguments{
\item{sequence}{a \code{character} object of length 1}
}
\value{
a character vector of DNA triplets

a character vector of DNA triplets

a character vector of DNA triplets

a character vector of DNA triplets
}
\description{
Codon Conversion

Codon Conversion

Codon Conversion
}
\examples{
orf <- Biostrings::DNAString("ATGAAATAG")
as_codons(orf)

rna <- Biostrings::RNAString("AUGAAAUAG")
as_codons(rna)

rna <- "AUGAAAUAG"
as_codons(rna)

dna <- "ATGAAATAG"
as_codons(dna)

}
