% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonopt-package.R
\docType{package}
\name{codonopt-package}
\alias{codonopt}
\alias{codonopt-package}
\title{codonopt: Codon Optimality Tools}
\description{
Functions for charactizing codon optimality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mfansler.github.io/codonopt/}
}

}
\author{
\strong{Maintainer}: Mervin Fansler \email{fanslerm@mskcc.org} (\href{https://orcid.org/0000-0002-4108-4218}{ORCID})

}
\keyword{internal}
