% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc.R
\name{csc_default}
\alias{csc_default}
\alias{csc_hela_endo}
\alias{csc_k562_orfome}
\alias{csc_k562_slam}
\alias{csc_rpe_endo}
\alias{csc_x293t_endo}
\alias{csc_x293t_orfome}
\title{Codon Stability Coefficients}
\usage{
csc_default()

csc_hela_endo()

csc_k562_orfome()

csc_k562_slam()

csc_rpe_endo()

csc_x293t_endo()

csc_x293t_orfome()
}
\value{
A named numeric vector.
\describe{
  \item{names}{Codons (DNA triplets)}
  \item{values}{CSCs for K562_ORFome}
}
}
\description{
Codon stability coefficients (CSCs) objects.
}
\examples{
csc_default()
csc_k562_orfome()
}
