% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_csc}
\alias{df_csc}
\title{Codon Stabilization Coefficients}
\format{
A data frame.
\describe{
\item{codon}{Codons (DNA triplets)}
\item{aa_code}{Amino acid 3-letter codes}
\item{X293T_endo}{CSC for X293T cells}
\item{HeLa_endo}{CSC for HeLa cells}
\item{RPE_endo}{CSC for RPE cells}
\item{X293T_ORFome}{CSC for X293T cells}
\item{K562_ORFome}{CSC for K562 cells}
\item{K562_SLAM}{CSC for K562 cells}
\item{genome}{Codon counts in genome}
\item{transcriptome}{Codon counts weighted by gene RPKM}
}
}
\source{
https://doi.org/10.7554/eLife.45396.006
}
\usage{
df_csc
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Table of codon stabilization coefficients (CSCs) as reported by
\url{https://doi.org/10.7554/eLife.45396}.
}
\keyword{datasets}
