% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{runiforf}
\alias{runiforf}
\title{Uniform Open Reading Frame Distribution}
\usage{
runiforf(sequence, size = 1)
}
\arguments{
\item{sequence}{a \code{Biostrings::AAString} or translatable \code{Biostrings::DNAString}
or \code{Biostrings::RNAString}}

\item{size}{integer number of samples to take}
}
\value{
a character vector of sampled open reading frames
}
\description{
Implements a uniform distribution of possible open reading frames
conditioned on an amino acid sequence or open reading frame.
}
