% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc.R
\name{.get_csc}
\alias{.get_csc}
\title{Extract specific CSC data}
\usage{
.get_csc(assay = "K562_SLAM")
}
\arguments{
\item{assay}{character of experiment code}
}
\value{
A named numeric vector.
}
\description{
Extract specific CSC data
}
\seealso{
df_csc

csc_default
}
