% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_codons}
\alias{as_codons}
\alias{as_codons,DNAString-method}
\alias{as_codons,RNAString-method}
\alias{as_codons,character-method}
\title{Codon Conversion}
\usage{
as_codons(sequence)

\S4method{as_codons}{DNAString}(sequence)

\S4method{as_codons}{RNAString}(sequence)

\S4method{as_codons}{character}(sequence)
}
\arguments{
\item{sequence}{an object representing a translatable sequence}
}
\value{
a character vector of DNA triplets
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Method for converting from DNA or RNA strings in various formats
to a set of codons.
}
\details{
This converts a sequence to a character vector of DNA triplets.
This function is not vectorized. Supported input types are \code{DNAString},
\code{RNAString}, and \code{character} vector.
}
\examples{
orf <- Biostrings::DNAString("ATGAAATAG")
as_codons(orf)

rna <- Biostrings::RNAString("AUGAAAUAG")
as_codons(rna)

rna <- "AUGAAAUAG"
as_codons(rna)

dna <- "ATGAAATAG"
as_codons(dna)

}
