% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{score_mean_csc}
\alias{score_mean_csc}
\title{Score for Mean Codon Stability Coefficient}
\usage{
score_mean_csc(sequence, csc = csc_default())
}
\arguments{
\item{sequence}{a character vector with DNA triplets}

\item{csc}{a named numeric}
}
\value{
a numeric value in \link{0,1}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Computes the mean codon stability score for a sequence.
}
\examples{
codons <- c("ATG", "TTA", "GAT", "TTA", "TAG")
score_mean_csc(codons)

score_mean_csc(as_codons("AUGUUAGAUUUAUAG"))
}
